# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
			"IUsers": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"properties": {
								"manipolicity": {
									"type": "string"
								},
								"government": {
									"type": "string"
								},
								"country": {
									"type": "string"
								}
							},
							"required": [
								"manipolicity",
								"government",
								"country"
							],
							"type": "object"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSent",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSetting": {
						"additionalProperties": false,
						"type": "object"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"phoneNumber",
					"password",
					"profilePicture",
					"firstName",
					"lastName",
					"gender",
					"DOB",
					"verified",
					"status",
					"accountType",
					"lastLoginDate",
					"accountSetting",
					"languages",
					"maritalStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IDisputes": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							}
						]
					},
					"firstPartyUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"secondUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"properties": {
								"disType": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"disType",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"received",
							"in-progress",
							"suspended",
							"rejected",
							"resolved"
						]
					},
					"resolverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"inProgressDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"suspendedDate": {
						"type": "string",
						"format": "date-time"
					},
					"closedDate": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"topic",
					"details",
					"status",
					"inProgressDate",
					"receivedDate",
					"suspendedDate",
					"closedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IDisputes_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							}
						]
					},
					"firstPartyUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"secondUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"topic": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"properties": {
								"disType": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"disType",
								"name",
								"url"
							],
							"type": "object"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"sent",
							"received",
							"in-progress",
							"suspended",
							"rejected",
							"resolved"
						]
					},
					"resolverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"inProgressDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"suspendedDate": {
						"type": "string",
						"format": "date-time"
					},
					"closedDate": {
						"type": "string",
						"format": "date-time"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IAgreements": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"parties": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"type": "string"
					},
					"attachments": {
						"items": {
							"additionalProperties": false,
							"type": "object"
						},
						"type": "array"
					},
					"reminder": {
						"type": "boolean"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"parties",
					"startDate",
					"endDate",
					"details",
					"reminder"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IEmployees": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"roleID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IRoles"
							}
						]
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"attachments": {
						"items": {
							"properties": {
								"empType": {
									"type": "string"
								},
								"url": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"empType",
								"url",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"salary"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_IEmployees_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							}
						]
					},
					"uID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"roleID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IRoles"
							}
						]
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"attachments": {
						"items": {
							"properties": {
								"empType": {
									"type": "string"
								},
								"url": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"empType",
								"url",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"INotifications": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Agreement",
							"Dispute",
							"Alarm",
							"Announcement",
							"User",
							"Communication"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							},
							{
								"$ref": "#/components/schemas/IAgreements"
							},
							{
								"$ref": "#/components/schemas/IDisputes"
							},
							{
								"$ref": "#/components/schemas/IAlarms"
							},
							{
								"$ref": "#/components/schemas/IAnnouncements"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							},
							{
								"$ref": "#/components/schemas/ICommunications"
							}
						]
					},
					"statues": {
						"type": "string",
						"enum": [
							"sent",
							"delivered",
							"opened"
						]
					},
					"title": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"openDate": {
						"type": "string",
						"format": "date-time"
					},
					"receiverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"referenceType",
					"referenceID",
					"statues",
					"title",
					"details",
					"sentDate",
					"receivedDate",
					"openDate",
					"receiverUID"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_INotifications_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"referenceType": {
						"type": "string",
						"enum": [
							"Session",
							"Agreement",
							"Dispute",
							"Alarm",
							"Announcement",
							"User",
							"Communication"
						]
					},
					"referenceID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/ISessions"
							},
							{
								"$ref": "#/components/schemas/IAgreements"
							},
							{
								"$ref": "#/components/schemas/IDisputes"
							},
							{
								"$ref": "#/components/schemas/IAlarms"
							},
							{
								"$ref": "#/components/schemas/IAnnouncements"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							},
							{
								"$ref": "#/components/schemas/ICommunications"
							}
						]
					},
					"statues": {
						"type": "string",
						"enum": [
							"sent",
							"delivered",
							"opened"
						]
					},
					"title": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"sentDate": {
						"type": "string",
						"format": "date-time"
					},
					"receivedDate": {
						"type": "string",
						"format": "date-time"
					},
					"openDate": {
						"type": "string",
						"format": "date-time"
					},
					"receiverUID": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"$ref": "#/components/schemas/IUsers"
							}
						]
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/disputes/{disputeId}": {
			"get": {
				"operationId": "GetDispute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IDisputes"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested dispute is not found"
					}
				},
				"description": "Get a dispute details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "disputeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63107cdf144972d5a75c14cb"
					}
				]
			},
			"delete": {
				"operationId": "DeleteDispute",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IDisputes"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested dispute is not found"
					}
				},
				"description": "Delete a dispute",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "disputeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63107cdf144972d5a75c14cb"
					}
				]
			}
		},
		"/disputes/create": {
			"post": {
				"operationId": "CreateDispute",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IDisputes"
								},
								"examples": {
									"Example 1": {
										"value": {
											"sessionID": "63130e088d796320172e58de",
											"firstPartyUID": "63130ba28d796320172e58d6",
											"secondUID": "63124dffdcf1e4974079a441",
											"topic": "treatment",
											"details": "more details",
											"attachments": [
												{
													"url": "http://localhost:8000/disputes",
													"name": "file",
													"disType": "pdf"
												}
											],
											"status": "sent",
											"resolverUID": "63124dffdcf1e4974079a441",
											"inProgressDate": "2022-09-10T00:00:00.000Z",
											"receivedDate": "2022-09-10T00:00:00.000Z",
											"suspendedDate": "2022-09-10T00:00:00.000Z",
											"closedDate": "2022-09-10T00:00:00.000Z"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a dispute",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDisputes"
							}
						}
					}
				}
			}
		},
		"/disputes/update/{disputeId}": {
			"put": {
				"operationId": "UpdateDispute",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IDisputes"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a dispute",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "disputeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "63107cdf144972d5a75c14cb"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IDisputes_"
							},
							"example": "63107cdf144972d5a75c14cb"
						}
					}
				}
			}
		},
...
		"/employees/{employeeId}": {
			"get": {
				"operationId": "GetEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmployees"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested employee is not found"
					}
				},
				"description": "Get employee details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313151791b8bbcc188d8e37"
					}
				]
			},
			"delete": {
				"operationId": "DeleteEmployee",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmployees"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "The requested role is not found"
					}
				},
				"description": "Delete employee",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313151791b8bbcc188d8e37"
					}
				]
			}
		},
		"/employees/create": {
			"post": {
				"operationId": "CreateEmployee",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEmployees"
								},
								"examples": {
									"Example 1": {
										"value": {
											"uID": "63130ba28d796320172e58d6",
											"roleID": "631312818d796320172e58e5",
											"salary": 5000.5,
											"attachments": [
												{
													"name": "memo",
													"url": "www.google.com",
													"empType": "some type"
												}
											]
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation failed"
					}
				},
				"description": "Create employee",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmployees"
							}
						}
					}
				}
			}
		},
		"/employees/update/{employeeId}": {
			"put": {
				"operationId": "UpdateEmployee",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IEmployees"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update employee",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313151791b8bbcc188d8e37"
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IEmployees_"
							},
							"example": "6313151791b8bbcc188d8e37"
						}
					}
				}
			}
		},
...
		"/notifications/{notificationId}": {
			"get": {
				"operationId": "GetNotification",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/INotifications"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested notification in not found"
					}
				},
				"description": "Get a Notification details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteNotification",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/INotifications"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested notification in not found"
					}
				},
				"description": "Delete a notification",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6313410084cfa62399551059"
					}
				]
			}
		},
		"/notifications/create": {
			"post": {
				"operationId": "CreateNotification",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/INotifications"
								},
								"examples": {
									"Example 1": {
										"value": {
											"referenceType": "Session",
											"referenceID": "63131c361ea32f520cb28bd6",
											"statues": "opened",
											"title": "Warning",
											"details": "this notification contain on new details",
											"sentDate": "2022-09-10T00:00:00.000Z",
											"receivedDate": "2022-09-10T00:00:00.000Z",
											"openDate": "2022-09-10T00:00:00.000Z",
											"receiverUID": "63130ba28d796320172e58d6"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a notification",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/INotifications"
							}
						}
					}
				}
			}
		},
		"/notifications/update/{notificationId}": {
			"put": {
				"operationId": "UpdateNotification",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/INotifications"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a notification",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "notificationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_INotifications_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}